Finished dynamic menu generation based on available modules.
If the config file is incomplete, missing keys are filled in automatically.
If no config file is provided, the entire config file is guessed. This is
done based on the name of the directory.

The documentation for creating a new module are updated to instruct programmers
to include options for users to override their settings.