Remove obsolete publication type field
This field is used in datarequests of version youth-0. The field has been removed entirely in version youth-1. It's not worth keeping around for backward compatibility, because once this commit is deployed in production, there will only be a few youth-0 datarequests that aren't fully processed yet. For the processing of those requests, the publication type field is of no importance anymore (and if it is, it can still be viewed by looking at the full data request).