YDA-5651: extract parameter for .part uploads
Add a parameter so that temporarily using a .part filename for
files that are being uploaded in the portal can be disabled. This
can be necessary on systems where renaming large .part data objects
to their final name is very slow, such as S3 object storage systems
with the S3 storage plugin in consistent mode.